Extract AWSTGWRepository
